+++ kaparoo/filesystem/hierarchy/__init__.py
@@ -8,8 +8,10 @@ __all__ = (
     "Group",
     "Node",
     "Together",
+    "match",
 )
 
 from kaparoo.filesystem.hierarchy.base import Node
 from kaparoo.filesystem.hierarchy.entry import Directory, Entry, File
 from kaparoo.filesystem.hierarchy.group import Exclusive, Group, Together
+from kaparoo.filesystem.hierarchy.match import match